Fixed issues when the Soku2 engine is not loaded

+++ generate_build.sh
@@ -40,48 +40,52 @@ for character in `ls src/Characters/`; do
 	echo "Checking $character"
 	OUT="$OUTPUT/$character"
 	IN="src/Characters/$character"
-	FILES=""
-	mkdir -p "$OUT/data"
-	if [ -f "$OUT/$character.dat" ]; then
-		DATE1=$(stat -c '%Y' "$OUT/$character.dat")
-		for file in `find "$IN/data" -name "*.png" -o -name "*.xml" -o -name "*wav" -o -name "*.csv" -o -name "*.pal" -o -name "${character}_labels.json"`; do
-			DATE2=$(stat -c '%Y' "$file")
-			RESULT="$OUT/data/$(echo $file | tail -c +$(echo "$IN/data/" | wc -c) | sed 's/.json$/.txt/g')"
-			if [ $DATE1 -lt $DATE2 ] || ! [ -f $(echo $RESULT | sed 's/\.xml$/\.pat/g' | sed 's/\.csv$/\.cv1/g' | sed 's/\.png$/\.cv2/g' | sed 's/\.wav$/\.cv3/g') ]; then
+	if [ $(cat "$IN/character.json" | jq .enabled) '==' false ]; then
+		echo "Disabled"
+	else
+		FILES=""
+		mkdir -p "$OUT/data"
+		if [ -f "$OUT/$character.dat" ]; then
+			DATE1=$(stat -c '%Y' "$OUT/$character.dat")
+			for file in `find "$IN/data" -name "*.png" -o -name "*.xml" -o -name "*wav" -o -name "*.csv" -o -name "*.pal" -o -name "${character}_labels.json"`; do
+				DATE2=$(stat -c '%Y' "$file")
+				RESULT="$OUT/data/$(echo $file | tail -c +$(echo "$IN/data/" | wc -c) | sed 's/.json$/.txt/g')"
+				if [ $DATE1 -lt $DATE2 ] || ! [ -f $(echo $RESULT | sed 's/\.xml$/\.pat/g' | sed 's/\.csv$/\.cv1/g' | sed 's/\.png$/\.cv2/g' | sed 's/\.wav$/\.cv3/g') ]; then
+					mkdir -p $(dirname "$RESULT")
+					cp -f $file $RESULT
+					FILES="yes"
+				fi
+			done
+		else
+			for file in `find "$IN/data" -name "*.png" -o -name "*.xml" -o -name "*wav" -o -name "*.csv" -o -name "*.pal" -o -name "${character}_labels.json"`; do
+				RESULT="$OUT/data/$(echo $file | tail -c +$(echo "$IN/data/" | wc -c) | sed 's/.json$/.txt/g')"
 				mkdir -p $(dirname "$RESULT")
 				cp -f $file $RESULT
 				FILES="yes"
-			fi
-		done
-	else
-		for file in `find "$IN/data" -name "*.png" -o -name "*.xml" -o -name "*wav" -o -name "*.csv" -o -name "*.pal" -o -name "${character}_labels.json"`; do
-			RESULT="$OUT/data/$(echo $file | tail -c +$(echo "$IN/data/" | wc -c) | sed 's/.json$/.txt/g')"
-			mkdir -p $(dirname "$RESULT")
-			cp -f $file $RESULT
-			FILES="yes"
-		done
+			done
+		fi
+		if [ -z $FILES ]; then
+			echo "Assets are clean"
+		else
+			for file in `find "$OUT/data" -name "*.png" -o -name "*.xml" -o -name "*wav" -o -name "*.csv"`; do
+				cd "$(dirname $file)"
+				shady-cli convert $file | grep --color=never Converting
+				rm $file
+				cd - >/dev/null
+			done
+			rm -f "$OUTPUT/$character/$character.dat"
+			shady-cli pack -o "$OUT/$character.dat" -m data "$OUT"
+		fi
+		DLL="$(cat "$IN/character.json" | jq .character_dll | sed -r 's/^"(.*)"$/\1/g')"
+		echo "Building $DLL"
+		cmake --build $OUTPUT --target $(echo $DLL | sed -r 's/^(.*)\..*$/\1/g')
+		mkdir -p "$OUTPUT/Soku2_package/characters/$character" "$OUTPUT/standalone/characters/$character"
+		cp "$OUTPUT/src/Characters/$character/$DLL" "$OUT/$character.dat" "$IN/character.json" "$OUTPUT/standalone/characters/$character"
+		cp "$OUTPUT/src/Characters/$character/$DLL" "$OUT/$character.dat" "$IN/character.json" "$OUTPUT/Soku2_package/characters/$character"
+		echo ";#" > "$OUTPUT/Soku2_package/characters/$character/dummy.asm"
+		echo "return function() end" > "$OUTPUT/Soku2_package/characters/$character/dummy.lua"
+		cat "$IN/data/csv/$character/deck.csv" "$IN/data/csv/$character/deck.csv" "$IN/data/csv/$character/deck.csv" "$IN/data/csv/$character/deck.csv" > "$OUTPUT/Soku2_package/characters/$character/deck.cfg"
 	fi
-	if [ -z $FILES ]; then
-		echo "Assets are clean"
-	else
-		for file in `find "$OUT/data" -name "*.png" -o -name "*.xml" -o -name "*wav" -o -name "*.csv"`; do
-			cd "$(dirname $file)"
-			shady-cli convert $file | grep --color=never Converting
-			rm $file
-			cd - >/dev/null
-		done
-		rm -f "$OUTPUT/$character/$character.dat"
-		shady-cli pack -o "$OUT/$character.dat" -m data "$OUT"
-	fi
-	DLL="$(cat "$IN/character.json" | jq .character_dll | sed -r 's/^"(.*)"$/\1/g')"
-	echo "Building $DLL"
-	cmake --build $OUTPUT --target $(echo $DLL | sed -r 's/^(.*)\..*$/\1/g')
-	mkdir -p "$OUTPUT/Soku2_package/characters/$character" "$OUTPUT/standalone/characters/$character"
-	cp "$OUTPUT/src/Characters/$character/$DLL" "$OUT/$character.dat" "$IN/character.json" "$OUTPUT/standalone/characters/$character"
-	cp "$OUTPUT/src/Characters/$character/$DLL" "$OUT/$character.dat" "$IN/character.json" "$OUTPUT/Soku2_package/characters/$character"
-	echo ";#" > "$OUTPUT/Soku2_package/characters/$character/dummy.asm"
-	echo "return function() end" > "$OUTPUT/Soku2_package/characters/$character/dummy.lua"
-	cat "$IN/data/csv/$character/deck.csv" "$IN/data/csv/$character/deck.csv" "$IN/data/csv/$character/deck.csv" "$IN/data/csv/$character/deck.csv" > "$OUTPUT/Soku2_package/characters/$character/deck.cfg"
 done
 cd "$OUTPUT/Soku2_package"
 echo "Generating Soku2Addon.zip"
